管理員清單後端設置（B2E）

+++ app/controllers/API/ManagerController.php
@@ -9,17 +9,28 @@ class API_ManagerController extends \BaseController {
      */
     public function index()
     {
-        //
-    }
+        $managers = Manager::all();
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return Response
-     */
-    public function create()
-    {
-        //
+        foreach ($managers as $manager) {
+
+            $user = IltUser::find($manager->u_id);
+
+            $m = array();
+            $m['username']  = $user->username;
+            $m['name']      = $manager->m_name;
+            $m['title']     = $manager->m_title;
+            $m['status']    = $manager->m_status;
+
+            if ( Session::has('user.login') and (Session::get('user.is_sa') == true) ) {
+                $m['id']     = $manager->m_id;
+            }
+
+            $list = $m;
+        }
+
+        $response['managers'] = $list;
+        $response['status'] = '200 OK';
+        return Response::json($response);
     }
 
     /**
@@ -34,17 +45,25 @@ class API_ManagerController extends \BaseController {
         $validator  = Validator::make(Input::all(), $rules, $messages);
 
         if ($validator->fails()) {
-            return Response::json();
+            $messages = $validator->messages()->all();
+            $response['status'] = '400 Bad Request';
+            $response['msg'] = $messages;
+
+            return Response::json($response);
         }
         else {
-            $manager = new Manager;
-            $manager->u_id    = Input::get('u_id');
-            $manager->m_name  = Input::get('name');
+            $user        = IltUser::where('username', '=', Input::get('username'))->first();
+            $complainant = Complainant::where('u_id', '=', $user->u_id)->first();
+            $manager     = new Manager;
+
+            $manager->u_id    = $user->u_id;
+            $manager->m_name  = $complainant->name;
             $manager->m_title = Input::get('title');
-            $manager->m_email = Input::get('email');
+            $manager->m_email = $complainant->email;
             $manager->save();
 
-            return Response::json();
+            $response['status'] = '200 OK';
+            return Response::json($response);
         }
     }
 
@@ -56,18 +75,29 @@ class API_ManagerController extends \BaseController {
      */
     public function show($id)
     {
-        //
-    }
+        if (! ( Session::has('user.login') and (Session::get('user.is_sa') == true) )) {
+            $response['status'] = '403 Forbidden';
+            return Response::json($response);
+        }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return Response
-     */
-    public function edit($id)
-    {
-        //
+        $manager = Manager::find($id);
+
+        if ( null == $manager ) {
+            $response['status'] = '404 Not Found';
+            return Response::json($response);
+        }
+
+        $user = IltUser::find($manager->u_id);
+
+        $m = array();
+        $m['username']  = $user->username;
+        $m['name']      = $manager->m_name;
+        $m['title']     = $manager->m_title;
+        $m['status']    = $manager->m_status;
+
+        $response['managers'] = $m;
+        $response['status'] = '200 OK';
+        return Response::json($response);
     }
 
     /**
@@ -78,23 +108,51 @@ class API_ManagerController extends \BaseController {
      */
     public function update($id)
     {
-        $rules      = Config::get('vallidation.manager.update.rules');
-        $messages   = Config::get('vallidation.manager.update.massages');
+        $rules      = Config::get('vallidation.manager.update.type');
+        $messages   = Config::get('vallidation.manager.update.type');
         $validator  = Validator::make(Input::all(), $rules, $messages);
 
         if ($validator->fails()) {
-            return Response::json();
+            $messages = $validator->messages()->all();
+            $response['status'] = '400 Bad Request';
+            $response['msg'] = $messages;
+
+            return Response::json($response);
         }
-        else {
-            $manager = Manager::find($id);
-            $manager->u_id    = Input::get('u_id');
-            $manager->m_name  = Input::get('name');
-            $manager->m_title = Input::get('title');
-            $manager->m_email = Input::get('email');
+
+        $type = Input::get('type');
+
+        $rules      = Config::get('vallidation.manager.update.' . $type);
+        $messages   = Config::get('vallidation.manager.update.' . $type);
+        $validator  = Validator::make(Input::all(), $rules, $messages);
+
+        if ($validator->fails()) {
+            $messages = $validator->messages()->all();
+            $response['status'] = '400 Bad Request';
+            $response['msg'] = $messages;
+
+            return Response::json($response);
+        }
+
+        $manager = Manager::find($id);
+
+        if ( $type == 'files' ) {
+            $manager->title = Input::get('title');
+            $manager->save();
+        }
+        else if ( $type == 'status' ) {
+            $status = Input::get('status');
+
+            $manager->status = $status;
             $manager->save();
 
-            return Response::json();
+            $action = new Action;
+            $action->type   = ($status == 1) ? 'RECOVERY_MANAGER_PERM' : 'STOP_MANAGER_PERM';
+            $action->event  = Input::get('reason');
         }
+
+        $response['status'] = '200 OK';
+        return Response::json($response);
     }
 
     /**
@@ -105,8 +163,37 @@ class API_ManagerController extends \BaseController {
      */
     public function destroy($id)
     {
+        if ( Session::has('user.login') and (Session::get('user.is_sa') == true) ) {
+            $m['id']     = $manager->m_id;
+        }
+
+        $rules      = Config::get('vallidation.manager.delete');
+        $messages   = Config::get('vallidation.manager.delete');
+        $validator  = Validator::make(Input::all(), $rules, $messages);
+
+        if ($validator->fails()) {
+            $messages = $validator->messages()->all();
+            $response['status'] = '400 Bad Request';
+            $response['msg'] = $messages;
+
+            return Response::json($response);
+        }
+
+        $user   = IltUser::where('username', '=', Input::get('username'))->first();
         $manage = Manager::find($id);
-        $manager->delete();
+
+        if ( $id != $user->u_id or $manage->m_name != Input::get('name') ) {
+            $response['status'] = '400 Bad Request';
+            $response['msg'] = array('您輸入的資料與要求的目標不符！');
+
+            return Response::json($response);
+        }
+        else {
+            $manager->delete();
+
+            $response['status'] = '200 OK';
+            return Response::json($response);
+        }
     }
 
 }
